Add files via upload
This is the first Web Application (sorta?) which I made while learning JavaScript and its basics. Tech stack includes: HTML5, CSS3 and Vanilla JS.
A user can add todo items, remove todo items, and this will all be dealt with using the Local Storage, so, even if the user exits the browser, he/she will be able to access the previously entered items!